refactor(event): work in feedback for image ingestion
